requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azure-functions-durable
beautifulsoup4==4.13.4
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
clr_loader==0.2.7.post0
ConfigArgParse==1.7.1
curl_cffi==0.11.1
frozendict==2.4.6
html5lib==1.1
idna==3.10
llvmlite==0.44.0
lxml==5.4.0
multitasking==0.0.11
numba==0.61.2
numexpr==2.10.2
numpy==2.2.6
pandas==2.2.3
peewee==3.18.1
platformdirs==4.3.8
protobuf==6.31.1
pycparser==2.22
pyodbc==5.2.0
python-dateutil==2.9.0.post0
pythonnet==3.0.5
pytz==2025.2
requests==2.32.3
six==1.17.0
soupsieve==2.7
ta==0.11.0
tabulate==0.9.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
webencodings==0.5.1
websockets==15.0.1
yfinance==0.2.61
